PURCHASE ENDPOINT WORKS AND IT AWARDS XP

+++ banking/src/main/kotlin/com/project/banking/entities/XpTierEntity.kt
@@ -6,8 +6,6 @@ import jakarta.persistence.GeneratedValue
 import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 import jakarta.persistence.Table
-import java.math.BigDecimal
-
 
 @Entity
 @Table(name = "xp_tiers")
@@ -28,7 +26,7 @@ data class XpTierEntity (
      var name: String? = null,
 
     @Column(name = "xp_perk_multiplier", nullable = false)
-     var xpPerkMultiplier: BigDecimal? = null,
+     var xpPerkMultiplier: Double? = null,
 
     @Column(name = "xp_per_notification", nullable = false)
      var xpPerNotification: Long? = null,
