+++ common/src/main/kotlin/com/project/common/data/requests/xp/CreateXpTierRequest.kt
@@ -1,7 +1,6 @@
 package com.project.common.data.requests.xp
 
 import jakarta.validation.constraints.*
-import java.math.BigDecimal
 
 data class CreateXpTierRequest(
     @field:Min(0)
@@ -11,7 +10,7 @@ data class CreateXpTierRequest(
     @field:NotBlank
     val name: String,
     @field:Positive
-    val xpPerkMultiplier: BigDecimal,
+    val xpPerkMultiplier: Double,
     @field:Positive
     val xpPerNotification: Long,
     @field:Positive
